backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
ortools>=9.6.2534
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
